AiAgentGame2/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-socketio>=5.10.0
eventlet>=0.35.0
